удалила main.e2e.test
